used typed client (#25)
Provide a single location to configure and interact with a particular HttpClient. For example, a single typed client might be used:
- For a single backend endpoint.
- To encapsulate all logic dealing with the endpoint.
Work with DI and can be injected where required in the app.

Create an instance of HttpClient.
Create an instance of Token, passing in the instance of HttpClient to its constructor.